fix: `open_yazi_in_directory` error in nvim 0.10.1 & nightly
Fixes https://github.com/mikavilpas/yazi.nvim/issues/308
